Add a very dummy high-pass filter
I have coded the function `filter_noise`, that it is located in
`filter.jl`. It is a very simple high-pass filter that it is coded
in order to eliminate the very low frequencies spurious signal,
in particular, the one-over-f noise. It does not work for the
atmospheric noise.

Now the strip2 module, can create random sky representation based
on a given CMB power spectrum.

I have been start coding a map-maker. In order to resolve the
map-maker inverse problem, for TODs comphensive of noise sources,
I have developed the ConjGrad.jl Julia package, that is now
available on my personal github page.

As I thought, the map-makring is particulary tricky to code.
If I have time i will deploy the whole package, comphensive of
map-maker, in a few days.