Creation of Q and U maps
I have coded the routine `make_CMB_pol_maps` in order to create
Q, U, E and B maps starting from a given power-spectrum.

But there is a BUG in the clEE and ClBB computation. The routine
return and error when try to convolve the EE and BB power-spectrum
with the random realization.